update moved archives
