+++ app/src/main/java/com/inercy/hidroponia/MainActivity.kt
@@ -6,42 +6,12 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.annotation.RequiresApi
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import androidx.navigation.compose.currentBackStackEntryAsState
-import androidx.navigation.compose.rememberNavController
-import com.inercy.hidroponia.navigation.AppNavHost
-import com.inercy.hidroponia.navigation.AppTopBar
-import com.inercy.hidroponia.navigation.getTitleForRoute
-import com.inercy.hidroponia.ui.theme.HidroponiaTheme
 
 class MainActivity : ComponentActivity() {
     @RequiresApi(Build.VERSION_CODES.O)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContent {
-            HidroponiaTheme {
-                val navController = rememberNavController()
-                val backStackEntry by navController.currentBackStackEntryAsState()
-
-                Scaffold(
-                    topBar = {
-                        AppTopBar(
-                            title = getTitleForRoute(backStackEntry?.destination?.route),
-                            canNavigateBack = navController.previousBackStackEntry != null,
-                            navigateUp = { navController.navigateUp() }
-                        )
-                    }
-                ) { innerPadding ->
-                    AppNavHost(
-                        navController = navController,
-                        modifier = Modifier.padding(innerPadding)
-                    )
-                }
-            }
-        }
+        setContent { HidroponiaApp() }
     }
 }
