--- app/src/main/java/com/inercy/hidroponia/data/model/Status.kt
@@ -1,5 +1,0 @@
-package com.inercy.hidroponia.data.model
-
-enum class Status {
-    ACTIVO, INACTIVO
-}
