+++ app/src/main/java/com/inercy/hidroponia/domain/models/Enfermedad.kt
@@ -1,4 +1,4 @@
-package com.inercy.hidroponia.data.model
+package com.inercy.hidroponia.domain.models
 
 import android.media.Image
