+++ app/src/main/java/com/inercy/hidroponia/domain/models/Inventario.kt
@@ -1,4 +1,4 @@
-package com.inercy.hidroponia.data.model
+package com.inercy.hidroponia.domain.models
 
 data class Inventario(
     val nombre: String,
