+++ app/src/main/java/com/inercy/hidroponia/domain/models/Operador.kt
@@ -1,4 +1,4 @@
-package com.inercy.hidroponia.data.model
+package com.inercy.hidroponia.domain.models
 
 data class Operador(
     val usuario: Usuario,
