+++ app/src/main/java/com/inercy/hidroponia/domain/models/Parametro.kt
@@ -1,4 +1,4 @@
-package com.inercy.hidroponia.data.model
+package com.inercy.hidroponia.domain.models
 
 data class Parametro(
     val maximo: Double,
