+++ app/src/main/java/com/inercy/hidroponia/domain/models/Status.kt
@@ -1,0 +1,5 @@
+package com.inercy.hidroponia.domain.models
+
+enum class Status {
+    ACTIVO, INACTIVO
+}
