+++ app/src/main/java/com/inercy/hidroponia/utils/enums/StatusEnum.kt
@@ -1,4 +1,4 @@
-package com.inercy.hidroponia.ui.enums
+package com.inercy.hidroponia.utils.enums
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.CheckCircle
